Bug fix: derive label dtype from input label.
This prevents clipping uint32 to uint16. Default is to derive from image, which is uint16. Now set to derive from input label image, so that if it is uint32 this dtype is preserved. Either derive from reference image (which also inherits shape, etc) or pass dtype, depending on task function. Apply fix to all tasks that use NGIO derive.

+++ src/scmultiplex/fractal/shift_by_shift.py
@@ -190,7 +190,7 @@ def shift_by_shift(
         new_shifted_label_name = label_name_to_shift + "_shifted"
 
     new_label_container = moving_ome_zarr.derive_label(
-        name=new_shifted_label_name, overwrite=True
+        name=new_shifted_label_name, dtype=img_array.dtype, overwrite=True
     )
     new_label_container.set_array(shifted_img_array)
